fix(pkg/controller/utils_test.go,pkg/sshd/server_test.go): add tests for fingerprint back
